requirements.txt
attrs==22.1.0
Automat==22.10.0
beautifulsoup4==4.11.1
camelot-py==0.10.1
certifi==2022.9.24
cffi==1.15.1
chardet==5.0.0
charset-normalizer==2.1.1
click==8.1.3
colorama==0.4.6
constantly==15.1.0
contourpy==1.0.6
cryptography==38.0.4
cssselect==1.2.0
cycler==0.11.0
decorator==5.1.1
distro==1.8.0
docopt==0.6.2
Elixir==0.7.1
et-xmlfile==1.1.0
filelock==3.8.0
fonttools==4.38.0
fuzzywuzzy==0.18.0
ghostscript==0.7
hyperlink==21.0.0
idna==3.4
incremental==22.10.0
itemadapter==0.7.0
itemloaders==1.0.6
Jinja2==3.1.2
jmespath==1.0.1
kiwisolver==1.4.4
Levenshtein==0.20.9
lxml==4.9.1
MarkupSafe==2.1.1
matplotlib==3.6.2
numpy==1.23.5
opencv-python==4.6.0.66
openpyxl==3.0.10
packaging==21.3
pandas==1.5.2
parsel==1.7.0
pbr==5.11.0
pdfminer.six==20221105
pdfquery==0.4.3
pdftopng==0.2.3
Pillow==9.3.0
pipreqs==0.4.11
Protego==0.2.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
PyDispatcher==2.0.6
pyOpenSSL==22.1.0
pyparsing==3.0.9
pypdf2==2.11.2
pyquery==1.4.3
python-dateutil==2.8.2
python-Levenshtein==0.20.9
pytz==2022.6
PyYAML==6.0
queuelib==1.6.2
rapidfuzz==2.13.7
regex==2022.10.31
requests==2.28.1
requests-file==1.5.1
roman==3.3
Scrapy==2.7.1
service-identity==21.1.0
six==1.16.0
soupsieve==2.3.2.post1
SQLAlchemy==0.7.10
sqlalchemy-migrate==0.11.0
sqlparse==0.4.3
tabula-py==2.6.0
tabulate==0.9.0
Tempita==0.5.2
textrazor==1.4.0
tldextract==3.4.0
Twisted==22.10.0
twisted-iocpsupport==1.0.2
typing-extensions==4.4.0
urllib3==1.26.13
w3lib==2.1.0
wget==3.2
xlrd==2.0.1
xlwt==0.7.2
yarg==0.1.9
zope.interface==5.5.2
